- refactored protocol.ts into multiple files - created single internal export that exposes client internal for tests - simplified exported api for client - removed references to nuid in favor of createInbox - changed test subscriptions that use callbacks to use iterators where appropiate.
