+++ src/mod.ts
@@ -1,13 +1,2 @@
-export {
-  ConnectionOptions,
-  ErrorCode,
-  Msg,
-  NatsConnection,
-  NatsError,
-  Nuid,
-  Payload,
-  ServersChanged,
-  Subscription,
-  Events,
-} from "../nats-base-client/mod.ts";
 export { connect } from "./connect.ts";
+export * from "../nats-base-client/mod.ts";
