+++ tests/helpers/mod.ts
@@ -2,4 +2,4 @@ export { check } from "./delay.ts";
 export { Lock } from "./lock.ts";
 export { TestServer, Connection } from "./test_server.ts";
 export { assertErrorCode } from "./asserts.ts";
-export { NatsServer } from "./launcher.ts";
+export { NatsServer, ServerSignals } from "./launcher.ts";
